move tests, format fixes
